DATAES-541: Fix ElasticsearchRestTemplateTests alias handling
This commit fixes:

* IllegalArgumentException on ElasticsearchRestTemplate#removeAlias
* NoSuchElementException on ElasticsearchRestTemplate#queryForAlias when
  no alias exists.
